все исправил в двух методах
утилитарность -
добавил final и пустой конструктор

имена -
массивы переименовал array1 и array2
переменную halfSumm убрал из кода
переменную medianIndex убрал
добавил пермеенную median

понятие медианы массива понято не верно -
исправил,сложил 2 медиальных значения и разделил
на два

перегрузка выполнена некорректно -
исправил ,в обоих методах double

lengthArr - неверный порядок слов -
поменял на arrLength

if - else исправил на тернарный оператор

Нельзя изменять состояние входных параметров-
сделал копии массивов

тесты работают

+++ src/test/java/com/nikita/homeworks/homework1/MedianTest.java
@@ -7,25 +7,25 @@ public class MedianTest {
 
     @Test
     public void testIntMedian() {
-        float result = Median.median(new int[] { 5, 5, 5, 100, 5, 5, 5, 5 });
+        double result = Median.median(new int[] { 5, 5, 5, 100, 5, 5, 5, 5 });
         Assert.assertEquals(5, result, 0);
     }
 
     @Test
     public void testIntMedianOddNumber() {
-        float result = Median.median(new int[] {1, 5, 2, 8, 7});
+       double result = Median.median(new int[] {1, 5, 2, 8, 7});
         Assert.assertEquals(5, result, 0);
     }
 
     @Test
     public void testIntMedianEvenNumber() {
-        float result = Median.median(new int[] {1, 6, 2, 8, 7, 2});
+        double result = Median.median(new int[] {1, 6, 2, 8, 7, 2});
         Assert.assertEquals(4, result, 0);
     }
 
     @Test
     public void testIntMedianEvenAverage() {
-        float result = Median.median(new int[] {1, 2, 3, 4});
+       double result = Median.median(new int[] {1, 2, 3, 4});
         Assert.assertEquals(2.5, result, 0);
     }
